requirements.txt
requests
elevenlabs
pydub
python-dotenv
fastapi
uvicorn
pydantic